\name{TurnsModels-package}
\alias{TurnsModels-package}
\alias{TurnsModels}
\docType{package}
\title{
\packageTitle{TurnsModels}
}
\description{
\packageDescription{TurnsModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{TurnsModels}
\packageIndices{TurnsModels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{TurnsModels}

Maintainer: \packageMaintainer{TurnsModels}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
